spring security

+++ src/main/java/com/example/demo/repository/AnimationRepository.java
@@ -3,5 +3,5 @@ package com.example.demo.repository;
 import com.example.demo.model.Animation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AnimationRepository extends JpaRepository<Animation, Long> {
+public interface AnimationRepository extends JpaRepository<Animation, Integer> {
 }
